+++ src/mainwindow.h
@@ -61,7 +61,7 @@ class MainWindow : public QMainWindow {
     int m_nMouseClick_X_Coordinate;
     int m_nMouseClick_Y_Coordinate;
     long long int currentTime = 0LL;
-    float speedFactor = 1.f;
+    double speedFactor = 1.0;
     double intervalRemainder = 0; // for speed factor
     const long long int INTERVAL = 200LL;
     const long long int SLIDER_RATIO = 1000LL;
